requirements.txt
## Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

## PDF Processing
pdfminer.six>=20221105

## Database & Storage
pymongo>=4.0.0
google-cloud-storage>=2.10.0

## Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
loguru>=0.7.0
aiofiles>=23.0.0

## Data Processing
PyYAML>=6.0
requests>=2.31.0

## Development & Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # for testing FastAPI endpoints